Address validation cleanup notes
Agent-Logs-Url: https://github.com/Vale717171/archive-of-oblivion/sessions/7b41ac62-49ec-4cdc-841b-ea0f13d6fc24

+++ lib/features/game/game_engine_provider.dart
@@ -3515,7 +3515,7 @@ class GameEngineNotifier extends AsyncNotifier<GameEngineState> {
   }
 
   String _selectHint(int level, List<String> hints) {
-    final index = (level - 1).clamp(0, hints.length - 1) as int;
+    final index = (level - 1).clamp(0, hints.length - 1);
     return 'Hint ${index + 1}/${hints.length}\n\n${hints[index]}';
   }
